Move to `debug terraform` command
