+++ cmd/bundle/bundle.go
@@ -24,8 +24,8 @@ func New() *cobra.Command {
 	cmd.AddCommand(newValidateCommand())
 	cmd.AddCommand(newInitCommand())
 	cmd.AddCommand(newSummaryCommand())
-	cmd.AddCommand(newDependenciesCommand())
 	cmd.AddCommand(newGenerateCommand())
+	cmd.AddCommand(newDebugCommand())
 	cmd.AddCommand(deployment.NewDeploymentCommand())
 	return cmd
 }
